SkillMatchBE/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
python-docx==1.2.0
pdfplumber==0.11.7
requests==2.32.5
openai==2.1.0
python-dotenv==1.1.1
faiss-cpu==1.12.0
sentence-transformers==5.1.1
openpyxl==3.1.5
langchain==0.3.27
langchain-community==0.3.31
langchain_groq==0.3.8